requirements.txt
# Core training
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
trl>=0.8.0
peft>=0.10.0

# Unsloth for efficient training
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# Cloud support
google-cloud-storage>=2.10.0

# Utilities
accelerate>=0.27.0
bitsandbytes>=0.43.0
